Remove old Use of Funds locale items
- Deleted the old test feature and renamed the new one

Ticket: https://phabricator.wikimedia.org/T384705
