+++ test/banners/desktop/C25_WMDE_Desktop_DE_00/components/BannerVar.spec.ts
@@ -6,7 +6,7 @@ import { newDynamicContent } from '@test/banners/dynamicCampaignContent';
 import { useOfFundsContent } from '@test/banners/useOfFundsContent';
 import { formItems } from '@test/banners/formItems';
 import { CurrencyEn } from '@src/utils/DynamicContent/formatters/CurrencyEn';
-import { desktopUseOfFundsFeatures } from '@test/features/UseOfFunds2024';
+import { desktopUseOfFundsFeatures } from '@test/features/UseOfFunds';
 import {
 	bannerContentAnimatedTextFeatures,
 	bannerContentDateAndTimeFeatures,
